requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
Flask==2.1.0
fonttools==4.52.4
future==1.0.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
latexcodec==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
monty==2022.9.9
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.2
packaging==24.0
palettable==3.3.3
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
pybtex==0.24.0
pymatgen==2023.9.25
pymongo==4.7.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.13.1
six==1.16.0
spglib==2.0.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.3.0
tqdm==4.64.1
tzdata==2024.1
uncertainties==3.1.7
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.0
PyMySQL~=1.1.1